Moving shared logic into common file
